refactor: timer 파일명 변경
