+++ src/pages/TimerPage/TimerPage.tsx
@@ -14,12 +14,13 @@ import { splitTasksByCompletion } from '@/shared/utils/timer';
 
 import HamburgerIcon from '@/shared/assets/svgs/btn_hamburger.svg?react';
 
-import FriendInfoCarousel from '@/components/molecules/FriendInfoCarousel';
-import Timer from '@/components/molecules/Timer';
-import TimerSideBar from '@/components/molecules/TimerSideBar';
-import TimerSideBox from '@/components/molecules/TimerSideBoxTemporary';
-import TimerTitle from '@/components/molecules/TimerTitle';
 import TimerPageTemplates from '@/components/templates/TimerPageTemplates';
+
+import FriendInfoCarousel from './components/CarouselFriendInfo';
+import Timer from './components/Timer';
+import TimerSideBar from './components/TimerSideBar';
+import TimerSideBox from './components/TimerSideBoxTemporary';
+import TimerTitle from './components/TimerTitle';
 
 dayjs.extend(utc);
 dayjs.extend(timezone);
@@ -76,8 +77,6 @@ const TimerPage: React.FC = () => {
 		}
 	}, [todos, selectedTodo]);
 
-	console.log(todos);
-
 	const { data: setData } = useGetMoribSet(selectedTodo || 0);
 
 	const urls = useMemo(() => setData?.data.map((item: MoribSetData) => item.url.trim()) || [], [setData]);
@@ -100,8 +99,6 @@ const TimerPage: React.FC = () => {
 
 	if (isLoading) return <div>Loading...</div>;
 	if (error) return <div>Error loading todos</div>;
-
-	console.log(targetCategoryName);
 
 	return (
 		<TimerPageTemplates>
