+++ src/pages/TimerPage/components/TimerFavicon.tsx
