+++ src/pages/TimerPage/components/TimerSideBoxTemporary.tsx
@@ -1,7 +1,7 @@
-import SideBox from '@/components/atoms/SideBox';
-import TimerFavicon from '@/components/atoms/TimerFavicon';
-
 import AddBtnIcon from '@/shared/assets/svgs/btn_add.svg?react';
+
+import TimerFavicon from '@/pages/TimerPage/components/TimerFavicon';
+import SideBox from '@/pages/TimerPage/components/TimerSideBox';
 import { Favicon_DATA } from '@/shared/mocks/faviconData';
 
 const TimerSideBox = () => {
